python/requirements.txt
# Working versions tested with Python 3.13
fastapi==0.116.1
uvicorn==0.35.0
google-generativeai==0.8.5
python-dotenv==1.1.1
pydantic==2.11.7
httpx==0.28.1